Implementation of fillAffectsTable and Next relations (#58)
* Implement parsing of call statements and pkb stubs for addUsesP, addModifiesP, addPatternIf, addPatternWhile and addCalls.

* Clean up code in accordance with review #49.

* Add addCall for stmt, addCalls for proc-proc relations, addReadVarTable for stmt-var and addPrintVarTable for stmt-var in both PKB and SP parser side, as well as addCallsT in PKB.

* Add cfg nodes.

* Implement adding to Next relations from SP parser.

* Update #include header.

* Add test case for Next relations.

* Remove useless files.

* Partial implementation of fillAffectsTable.

* Add Cfg data structure and API for DE and SourceParser to use.

* Modifications to the data type of params and return statements of CFG related functions from std::string to int.

* Improve logic and abstraction for adding to CFG and populating Next relations. Clean up code and fix bugs in SP Parser for keeping track of statements to add to CFG.

* Add functions to keep track of procedure statement number range.

* Implement code for filling Affects relations.

* Fix bugs in fillAffectsTable() for Affects relations, add integration test cases for CFG and affects relations, cleaned up code and resolved test case for parsing of multiple procedures.

* Reduce unnecessary DFS calls by keeping track of all valid assign stmt 2s and doing a single DFS for each assign stmt 1.

* Change data structure for the DFS from deque to stack.

* Add integration test case for Affects relation and fix fillAffectsTable() bug wrt ignoring modifications to target variable in a procedure call statement on a path.

* Update PKB API (#60)

* Update PKB data structure

* Update extractDesignAbstractions functions call order

* Update PKB API

* Fix broken unit test cases

* Fix broken integration test cases

* Add integration test cases for Affects and Next relations.

* Add integration test case for Affects relations and restructure test cases to fix stack memory issues.

* Update fillsAffectsTable Aglo (#61)

* Update PKB data structure

* Update extractDesignAbstractions functions call order

* Update PKB API

* Fix broken unit test cases

* Fix broken integration test cases

* Update fillsAffectTable algorithm

* Update PKB API

* Update variable names

* Add const keyword

* Update PKB API name

* Add integration test case for Affects relation.

* Fix bug in filling of Affects relations.

* Remove debugging code from fillAffectsTable().

* Add integration test cases for Next relations.

* Rename TestSimpleParser.cpp to TestSimpleParsing.cpp.

* Remove commented code as per review #58.

+++ Team26/Code26/src/spa/src/utils/Table.cpp
@@ -247,7 +247,7 @@ void Table::innerJoin(const Table& otherTable, const std::vector<std::pair<int, 
       std::vector<Row> rows = hashmap.at(key);
       for (const Row& row : rows) { // do cross product on relevant rows
         Row newRow = row;
-        for (int i = 0; i < otherHeader.size(); i++) {
+        for (size_t i = 0; i < otherHeader.size(); i++) {
           bool isInToDrop = std::find(droppedSecondIndexes.begin(), droppedSecondIndexes.end(), i) != droppedSecondIndexes.end();
           if (!isInToDrop) {
             newRow.emplace_back(otherRow[i]);
